refinery/flow-requirements.txt
# Flow CLI Requirements - Screen Tracking & Screenshot Collection

# Core dependencies
python-dotenv>=1.0.0
click>=8.1.7
markdown>=3.5.1

# Screenshot and image processing
pillow>=10.1.0
pyscreenshot>=3.1
pyautogui>=0.9.54

# OCR processing
# macOS: Apple Vision (native, high quality, no external install)
pyobjc-framework-Vision>=10.0; sys_platform == 'darwin'
# Linux/Windows: Tesseract fallback
pytesseract>=0.3.10; sys_platform != 'darwin'
opencv-python>=4.8.1.78; sys_platform != 'darwin'

# Data processing
pandas>=2.1.4
numpy>=1.24.4

# ChromaDB client for storing screenshots
chromadb>=0.4.22
pydantic>=2.5.0
httpx>=0.26.0

# Logging and monitoring
structlog>=23.2.0
psutil>=5.9.6
